# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CrearTienda:
    properties:
      address:
        example: Av. Principal 123
        type: string
      name:
        example: Tienda Central
        type: string
      phone:
        example: 555-0123
        type: string
    type: object
...
  handlers.Tienda:
    description: Modelo de tienda para la API
    properties:
      address:
        example: Av. Principal 123
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Tienda Central
        type: string
      phone:
        example: 555-0123
        type: string
    type: object
  handlers.TiendaDetalle:
    properties:
      activo:
        example: true
        type: boolean
      address:
        example: Av. Principal 123
        type: string
      created_at:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Tienda Central
        type: string
      phone:
        example: 555-0123
        type: string
      updated_at:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API para gestión de productos
  title: API de Productos
  version: "1.0"
paths:
...
  /ActivarDesactivarTiendas:
    patch:
      consumes:
      - application/json
      description: Cambia el estado activo/inactivo de una tienda
      parameters:
      - description: ID de la tienda
        in: query
        name: id
        required: true
        type: string
      - description: true para activar, false para desactivar
        in: query
        name: activate
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TiendaDetalle'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Activar/Desactivar tienda
      tags:
      - tiendas
...
  /ActualizarTiendas:
    put:
      consumes:
      - application/json
      description: Actualiza los datos de una tienda existente
      parameters:
      - description: ID de la tienda
        in: query
        name: id
        required: true
        type: string
      - description: Datos actualizados de la tienda
        in: body
        name: tienda
        required: true
        schema:
          $ref: '#/definitions/handlers.CrearTienda'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TiendaDetalle'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Actualizar tienda
      tags:
      - tiendas
...
  /CrearTiendas:
    post:
      consumes:
      - application/json
      description: Crea una nueva tienda en el sistema
      parameters:
      - description: Datos de la tienda
        in: body
        name: tienda
        required: true
        schema:
          $ref: '#/definitions/handlers.CrearTienda'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.TiendaDetalle'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear nueva tienda
      tags:
      - tiendas
...
  /EliminarTiendas:
    delete:
      consumes:
      - application/json
      description: Elimina una tienda del sistema
      parameters:
      - description: ID de la tienda
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Eliminar tienda
      tags:
      - tiendas
...
  /ListarTiendas:
    get:
      consumes:
      - application/json
      description: Obtiene la lista de todas las tiendas activas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Tienda'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar todas las tiendas
      tags:
      - tiendas
...
  /ObtenerTiendas:
    get:
      consumes:
      - application/json
      description: Obtiene los detalles de una tienda específica
      parameters:
      - description: ID de la tienda
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TiendaDetalle'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener tienda por ID
      tags:
      - tiendas
...
schemes:
- http
swagger: "2.0"